feat: student dashboard overview and course

+++ components/DashboardLayout.tsx
@@ -1,27 +1,88 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useMemo, useRef, useState } from "react";
 
 import { useRouter } from "next/router";
-import { Layout, Menu, Avatar, Dropdown } from "antd";
 import {
+  Layout,
+  Menu,
+  Avatar,
+  Dropdown,
+  Row,
+  Col,
+  Badge,
+  Tabs,
+  message,
+  Spin,
+  List,
+  Space,
+  Button,
+  notification,
+} from "antd";
+import {
+  BellOutlined,
   MenuUnfoldOutlined,
   MenuFoldOutlined,
   UserOutlined,
   LogoutOutlined,
 } from "@ant-design/icons";
+import styled from "styled-components";
 
 import { routes, SideNav } from "../lib/constant/routes";
 import { generateKey, getActiveKey } from "../lib/util/side-nav";
 
-import { logout } from "../lib/services/api-service";
+import {
+  getMessages,
+  getMessageStatistic,
+  logout,
+  markAsRead,
+  messageEvent,
+} from "../lib/services/api-service";
 import Link from "next/link";
 
 import SubMenu from "antd/lib/menu/SubMenu";
 import NavBreadcrumb from "./NavBreadcrumb";
-import { Role } from "../lib/model";
+import { Message, MessageType, Paginator, Role } from "../lib/model";
 import { useUserRole } from "./custom-hooks/Login-state";
 import storage from "../lib/services/storage";
+import { useMsgStatistic } from "./MessagesProvider";
+import InfiniteScroll from "react-infinite-scroll-component";
+import { formatDistanceToNow } from "date-fns";
 
 const { Header, Sider, Content } = Layout;
+
+const TabNavContainer = styled.div`
+  margin-bottom: 0;
+  padding: 10px 20px 0 20px;
+  .ant-tabs-nav-list {
+    width: 100%;
+    justify-content: space-around;
+  }
+`;
+
+const Footer = styled(Row)`
+  height: 50px;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  border-radius: 0 0 4px 4px;
+  border: 1px solid #f0f0f0;
+  border-left: none;
+  border-right: none;
+  background: #fff;
+  z-index: 9;
+  .ant-col {
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    &:first-child {
+      box-shadow: 1px 0 0 0 #f0f0f0;
+    }
+  }
+  button {
+    border: none;
+  }
+`;
 
 const getMenuConfig = (
   data: SideNav[],
@@ -72,6 +133,256 @@ const renderMenuItems = (
   });
 };
 
+interface MessageProps {
+  type: MessageType;
+  NewMessage?: Message;
+}
+
+const Messages = ({ type, NewMessage }: MessageProps) => {
+  const [messageList, setMessageList] = useState<Message[]>([]);
+  const [hasMore, setHasMore] = useState<boolean>(true);
+  const [total, setTotal] = useState<number>(0);
+  const [pagination, setPagination] = useState<Paginator>({
+    page: 1,
+    limit: 20,
+  });
+
+  const { dispatch } = useMsgStatistic();
+
+  useEffect(() => {
+    const req = {
+      page: pagination.page,
+      limit: pagination.limit,
+      userId: storage.userId,
+      type: type,
+    };
+    getMessages(req).then((resp) => {
+      if (resp) {
+        setTotal(resp.total);
+        setMessageList((pre) => [...pre, ...resp.messages]);
+      }
+    });
+  }, [pagination, type]);
+
+  useEffect(() => {
+    if (!!NewMessage && NewMessage.type === type) {
+      setMessageList((pre) => [NewMessage, ...pre]);
+    }
+  }, [NewMessage, type]);
+
+  return (
+    <>
+      <div
+        id={`msg-container-${type}`}
+        style={{ marginLeft: "20px", overflowY: "scroll", height: "380px" }}
+      >
+        <InfiniteScroll
+          next={() => {
+            setPagination({ ...pagination, page: pagination.page + 1 });
+            setHasMore(messageList.length < total);
+          }}
+          hasMore={hasMore}
+          loader={
+            <div style={{ textAlign: "center" }}>
+              <Spin />
+            </div>
+          }
+          dataLength={messageList.length}
+          endMessage={<div style={{ textAlign: "center" }}>No more</div>}
+          scrollableTarget={`msg-container-${type}`}
+        >
+          <List
+            itemLayout="vertical"
+            dataSource={messageList}
+            renderItem={(item, index) => (
+              <List.Item
+                key={index}
+                style={{ opacity: item.status ? 0.4 : 1 }}
+                actions={[
+                  <Space key={index}>
+                    {formatDistanceToNow(new Date(item.createdAt), {
+                      addSuffix: true,
+                    })}
+                  </Space>,
+                ]}
+                onClick={() => {
+                  if (item.status === 1) {
+                    return;
+                  }
+
+                  markAsRead([item.id]).then((res) => {
+                    if (res) {
+                      const target = messageList.find(
+                        (msg) => item.id === msg.id
+                      );
+
+                      target.status = 1;
+                      setMessageList([...messageList]);
+                    }
+
+                    dispatch({
+                      type: "decrement",
+                      payload: { count: 1, type: item.type },
+                    });
+                  });
+                }}
+              >
+                <List.Item.Meta
+                  avatar={<Avatar icon={<UserOutlined />} />}
+                  title={item.from.nickname}
+                  description={item.content}
+                />
+              </List.Item>
+            )}
+          ></List>
+        </InfiniteScroll>
+      </div>
+      <Footer>
+        <Col span={12}>
+          <Button
+            onClick={() => {
+              const ids = messageList
+                .filter((item) => item.status === 0)
+                .map((item) => item.id);
+
+              if (ids.length) {
+                markAsRead(ids).then((res) => {
+                  if (res) {
+                    setMessageList(
+                      messageList.map((item) => ({ ...item, status: 1 }))
+                    );
+                  }
+
+                  dispatch({
+                    type: "decrement",
+                    payload: { count: ids.length, type },
+                  });
+                });
+              } else {
+                message.warn(`All of these ${type}s has been marked as read!`);
+              }
+            }}
+          >
+            Mark all as read
+          </Button>
+        </Col>
+
+        <Col span={12}>
+          <Button>
+            <Link href={`/dashboard/${storage.role}/message`}>
+              View history
+            </Link>
+          </Button>
+        </Col>
+      </Footer>
+    </>
+  );
+};
+
+const MessagePanel = () => {
+  const types: MessageType[] = ["notification", "message"];
+  const { msgStore, dispatch } = useMsgStatistic();
+  const [activeType, setActiveType] = useState<MessageType>("notification");
+  const [message, setMessage] = useState<Message>(null);
+
+  useEffect(() => {
+    getMessageStatistic().then((resp) => {
+      if (!!resp) {
+        const { notification, message } = resp.receive;
+
+        dispatch({
+          type: "increment",
+          payload: { type: "message", count: message.unread },
+        });
+        dispatch({
+          type: "increment",
+          payload: { type: "notification", count: notification.unread },
+        });
+      }
+    });
+
+    const sse = new EventSource(
+      `https://cms.chtoma.com/api/message/subscribe?userId=${storage.userId}`,
+      {
+        withCredentials: true,
+      }
+    );
+
+    sse.onmessage = (event) => {
+      let { data } = event;
+
+      data = JSON.parse(data || {});
+
+      if (data.type !== "heartbeat") {
+        const content = data.content as Message;
+
+        if (content.type === "message") {
+          notification.info({
+            message: `You have a message from ${content.from.nickname}`,
+            description: content.content,
+          });
+        }
+
+        setMessage(content);
+        dispatch({
+          type: "increment",
+          payload: { type: content.type, count: 1 },
+        });
+      }
+    };
+
+    return () => {
+      sse.close();
+      dispatch({ type: "reset" });
+    };
+  }, [dispatch]);
+
+  return (
+    <Badge size="small" count={msgStore.total} offset={[10, 0]}>
+      <Dropdown
+        placement="bottomLeft"
+        trigger={["click"]}
+        overlayStyle={{
+          background: "#fff",
+          borderRadius: 4,
+          width: 400,
+          height: 500,
+          overflow: "hidden",
+        }}
+        overlay={
+          <>
+            <Tabs
+              animated
+              onChange={(key: MessageType) => {
+                if (key !== activeType) {
+                  setActiveType(key);
+                }
+              }}
+              renderTabBar={(props, DefaultTabBar) => {
+                return (
+                  <TabNavContainer>
+                    <DefaultTabBar {...props} />
+                  </TabNavContainer>
+                );
+              }}
+            >
+              {types.map((type) => {
+                return (
+                  <Tabs.TabPane tab={`${type} (${msgStore[type]})`} key={type}>
+                    <Messages type={type} NewMessage={message} />
+                  </Tabs.TabPane>
+                );
+              })}
+            </Tabs>
+          </>
+        }
+      >
+        <BellOutlined style={{ fontSize: 24, color: "white" }} />
+      </Dropdown>
+    </Badge>
+  );
+};
+
 const DashboardLayout = (props: React.PropsWithChildren<any>) => {
   const { children } = props;
   const router = useRouter();
@@ -87,18 +398,19 @@ const DashboardLayout = (props: React.PropsWithChildren<any>) => {
 
   const DropdownAvatar = () => {
     return (
-      <Dropdown overlay={menu} placement="bottomLeft">
+      <Dropdown overlay={avatarMenu} placement="bottomLeft">
         <Avatar style={{ marginRight: 20 }} icon={<UserOutlined />} />
       </Dropdown>
     );
   };
   const logoutHandler = async () => {
-    logout();
-    storage.deleteUserInfo();
-    router.push("/login");
+    logout().then((resp) => {
+      if (resp) storage.deleteUserInfo();
+      router.push("/login");
+    });
   };
 
-  const menu = (
+  const avatarMenu = (
     <Menu>
       <Menu.Item key="100" onClick={logoutHandler} icon={<LogoutOutlined />}>
         Logout
@@ -172,7 +484,14 @@ const DashboardLayout = (props: React.PropsWithChildren<any>) => {
               },
             }
           )}
-          <DropdownAvatar />
+          <Row gutter={[35, 16]} align="middle">
+            <Col style={{ top: 5 }}>
+              <MessagePanel />
+            </Col>
+            <Col>
+              <DropdownAvatar />
+            </Col>
+          </Row>
         </Header>
 
         <Content
